apply new norme

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: apaula-b <apaula-b@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/02/02 19:22:17 by apaula-b          #+#    #+#             */
-/*   Updated: 2021/02/28 11:40:38 by apaula-b         ###   ########.fr       */
+/*   Updated: 2021/05/01 13:00:32 by apaula-b         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(char *s)
 {
-	int size;
+	size_t	size;
 
 	size = 0;
 	while (s[size])
